# feat: optimize BirdNET prediction pipeline